+++ lib/springcm-sdk/attribute.rb
@@ -1,0 +1,17 @@
+require "springcm-sdk/resource"
+
+module Springcm
+  class Attribute < Resource
+    def required?
+      required
+    end
+
+    def read_only?
+      read_only
+    end
+
+    def repeating_attribute?
+      repeating_attribute
+    end
+  end
+end