+++ lib/springcm-sdk/mixins/attributes.rb
@@ -1,0 +1,4 @@
+module Springcm
+  module Attributes
+  end
+end